Refactor: Modularize design-system.ts and consolidate styles
This commit breaks down the monolithic `design-system.ts` into multiple, more focused modules under `src/lib/styles/`:
- `common.ts`: For common utility styles like focus rings, disabled states, and glassmorphism.
- `theme.ts`: For theme configurations, ThemeProvider, and useTheme hook.
- `components.ts`: For component-specific style objects (button, input, card, bento).
- `animations.ts`: For animation variants and transitions.
- `utils.ts`: For utility functions related to the design system.

The original `design-system.ts` now acts as an aggregator, re-exporting from these new modules to maintain backward compatibility for most imports and the default export.

Key changes include:
- Creation of the new style modules.
- Migration of relevant code from `design-system.ts` to these modules.
- Update of `design-system.ts` to re-export from the new modules.
- Update of import statements across the codebase to point to the new, more specific modules where appropriate.
- Consolidation of redundant style definitions, particularly by refactoring UI components in `src/components/ui/` (like card, input, bento) to use styles from `src/lib/styles/components.ts`.
- Cleanup of `src/lib/ui-styles.ts` by removing definitions now duplicated in `src/lib/styles/common.ts`.

Automated tests timed out during execution, so I recommend thorough manual testing to ensure no regressions were introduced.
