+++ src/components/ui/command.tsx
@@ -5,7 +5,7 @@ import { Search } from "lucide-react"
 
 import { cn } from "@/lib/utils"
 import { Dialog, DialogContent } from "@/components/ui/dialog"
-import { glass } from "@/lib/ui-styles"
+import { glass } from "@/lib/styles/common" // Updated import
 
 const Command = React.forwardRef<
   React.ElementRef<typeof CommandPrimitive>,
